+++ src/app/(main)/post/loading.tsx
@@ -1,5 +1,5 @@
-import PageSkeleton from '@/components/PageSkeleton';
+import PostPageSkeleton from '@/components/post/PostPageSkeleton';
 
-const Loading = () => <PageSkeleton />;
+const Loading = () => <PostPageSkeleton />;
 
 export default Loading;